Additional getters and setters
To avoid memory allocation when using this class as a wrapper over an underlying buffer it would be convenient to be able reposition the buffer which would be enabled with additional setters. 
Further, the getCapacity() is named after the Java implementation. Would it not be more correct to say getLength() or length() instead?

Thanks