[Java] Avoid processing "early" catch-up requests.
Details of the issue:
=====================

1. Three nodes were started at roughly the same time
2. Only one node had the "history" of log
3. This node with "history" was elected the leader
4. The leader entered the `LEADER_REPLAY` state and didn't immediately transition out
because `shouldReplay == true`
5. The followers published their catch-up positions
6. The leader polled for member status which dispatches catch-up
positions to the `onCatchupPosition(...)` handler
7. The leader dropped the catch-up positions because its leadership term
and role are not updated until `LEADER_TRANSITION` is reached.
Therefore, the log is never replayed to followers.

Fix in progress:
================

I've tried to skip polling for member status whilst the consensus module
for a node is in `{ LEADER_REPLAY, LEADER_TRANSITION }`. This means we
don't process catch-up positions (amongst other messages) from followers
during this leadership transition.

Big/scary assumptions:
======================

`LEADER_REPLAY` always transitions to `LEADER_READY`. If not, then I've
introduced the possibility for a node to become "stuck" in a
transitioning state.

Questions:
==========

1. Would it be better to signal `LEADER_READY` to followers before they
publish their catch-up positions?

2. Would it be better to transition the `role` and `leadershipTermId`
earlier rather than pausing member status processing?

3. Is the `LEADER_REPLAY` behaviour different on startup in order to
replay the log to the clustered service container? I.e., when it is
replaying is it replaying to "itself"?

Known issues with this change:
==============================

`ClusterTest.shouldCatchupFromEmptyLogThenSnapshotAfterShutdownAndFollowerCleanStart()`
now fails ~ 50% of the time. It fails because when processing an
acknowledgement of `Snapshot` from the clustered service container it
cannot find the entry for the previous leadership term in the recording
log. I believe this race was already present but I have made it worse;
however, I'm not 100% sure!