Add OSGi manifest headers
This change moves to using the biz.aQute.bnd Gradle plugin to
produce OSGi-compatible manifests in jar files. Currently, only the
aeron-client, aeron-driver, and aeron-cluster modules have been updated
to provide OSGi metadata because the other modules are samples, tests,
or should never need to be visible inside an OSGi container. Care
has been taken to preserve the existing defined manifest fields
(Automatic-Module-Name, Implementation-Title, etc).

Note that until Agrona 1.0.4 is released (or at least a version containing
https://github.com/real-logic/agrona/commit/49e769e07c3ba3fa176d3b1079f2472184612e1d),
the generated package imports for Agrona will be unversioned. This will
self-correct when Agrona 1.0.4+ is released and Aeron is updated to
point to the new version.