[C++] Fix narrowing conversion in client
MSCV is correctly complaining about the narrowing conversion from
`std::size_t` to `int` in `AtomicArrayUpdater::removeElement`.

This patch makes us return a `std::size_t` instead. Using `-1` to
signify an invalid length is not needed as the length value is not used
anyway if the pointer is null. Hence we use `0` as the invalid value and
avoid the narrowing conversion.