[Java] Re-implement replay using new `offerBlock` API
In order to apply some batching for the replay a new method `ExclusivePublication#offerBlock` was added. It is a low-level API intended for offering pre-formatted block of messages directly into publication. This method then copies provided block verbatim into the underlying term.
Preliminary performance tests showed a 50% increase in message throughput for the replay.