[Java] Use `AeronCloseHelper` with the `ErrorHandler` to close resources
Ensure proper closure of the resources by using new helper API that uses `ErrorHandler` which consumes
any error generated by the resource being closed. In cases where there is no `ErrorHandler` around or if
it was already closed use variants of the `CloseHelper#closeAll` methods and if not possible
`CloseHelper#quiteClose` as the final resort.

Because not all of the resource being closed implement the `AutoCloseable` interface we are forced to use
lambda expressions in several places. This might be a potential performance issues, however given the fact
that we are dealing with `close` methods that should never be on the fact path.