[Java] Fix the race between shutdown of the initial leader and addition of the new dynamic node
Test was only waiting for the new node to exit election. However the cluster change event
(i.e. `io.aeron.cluster.LogPublisher.appendMembershipChangeEvent`) might not have been called
by that time. Which means that cluster still had a size of 3.

The fix is to wait for the cluster to have size 4 before killing the old leader.