[C] Implement custom aeron_uuid for enable uuid on all platform.
The randombytes are using https://github.com/dsprenkels/randombytes for true random value generator.
The the style are not changed.

UUID is for generating a unique identifier for the computer, so that's important using a true random-generator across platform.
Remove the dependencies to libuuid make aeron to be a more portable message queue.
Other message system like nanomsg are also have zero dependencies.