[C] Do not fail if Aeron dir exists, but the CnC file does not.
If you try to start a C Media Driver with an existing directory,
but without a CnC file inside, it will fail:

ERROR: driver init (2) (2) No such file or directory
[aeron_open_file_rw, aeron_fileutil.c:428] Failed to open file: /dev/shm/empty/cnc.dat
[aeron_driver_init, aeron_driver.c:860] could not recreate aeron dir: /dev/shm/empty

ERROR: driver close (22) (22) Invalid argument
[aeron_driver_close, aeron_driver.c:1117] driver is null

It's not really an error, could be due to the directory being empty,
so it seems better to continue.

This aligns behaviour with the Java Media Driver.