[C] Fix spy subscription channel. (#1513)
When looking at AeronStat of a C media driver, noticed that a spy subscription
doesn't have the aeron-spy: prefix like in the Java driver. Turns out we
incorrectly copy the channel into the link - starting after the prefix. Also,
the copy uses the full length potentially trying to access data out of bound.
This was inconsistent with MDS code which is using the full channel for spies.

Image unavailability callbacks now get the full channel with prefix.