[Java] Expose counter for number of bytes retransmitted.
This commit adds a counter that tracks the number of bytes retransmitted
due to loss. The existing tests issue a variable number of NAKs
depending on timings. Locally with breakpoints set, I've seen > 30 NAKs.
Therefore, I have added assertions with only a lower bound based on the
number of dropped bytes.