[C] Expose counter for number of bytes retransmitted.
This commit adds a counter that tracks the number of bytes retransmitted
due to loss. The same counter has already been added to Java and the
tests have been extended.

It is worth noting that the counter value is only a lower bound as MDC
retransmits are only counted once regardless of the number of
destinations. Also, the retransmitted bytes are not included in the
total bytes sent counter.