[Java] Reduce archive listRecordingsForUri idling
With the recent reduction to MAX_SCANS_PER_WORK_CYCLE of list recordings, we noted that some list operations took significantly longer time, but when profiling it looked like archive was just sleeping.

As it turns out, list recordings only return a postive work count if it actually found something, but this means that a listRecordingsForUri API call that matches only a small subset of recordings (or even none at all) will idle every 64 recordings, which for a large catalog caused the list operation to take multiple seconds of sleeping.

This instead changes work count to be based on the number of recordings processed, which for listRecordings should be exactly the same, but for listRecordingsForUri should only idle if there are no more recordings to check at the moment.