C++ wrapper improvements (#1639)
* [C++] Fix Image wrapping in a couple of cases.

Both `aeron_subscription_image_by_session_id(m_subscription, sessionId)`
and `aeron_subscription_image_at_index(m_subscription, index)` may
return `NULL`. In this commit, I've fixed the wrapping code to handle
the null case.

* [C++] Delete implicit FragmentAssembler copy/move ctors/operators.

`m_fragment_assembler` is a C struct that captures a pointer to `this`
in the `FragmentAssembler` constructor. The default copy/move
constructors and operators are not safe, as the original `this` may have
been deleted by the time the copied/moved assembler is called.