[Java] Read high-water mark position after computing the rebuild position to ensure the widest possible loss detection scan window.
Loss detection code is concurrent to the application threads that are consuming from the image but also to the receiver thread that inserts new packets into the image and advances the high-water mark position counter. Reading that counter last ensures that we take into account all inserted packets up to this point.